requirements.txt
PyQt5
arabic-reshaper
python-bidi